Rename package name for OAuth2AuthenticationSessionDestroyedListener See gh-4866
